Refactor dependency resolution and resolver interfaces
I just realized the input to the dependency resolver will need to be
an instance of the dependency class.

Also, add kind and number to the dependency.

The number is a temporary field. Ideally, instead of a fixed number,
each dependency should have a list of requisites. Equality, for instance,
can be one of such requisites.