Introducing RepositoryFilter and RubyRepositoryFilter
RepositoryFilter is a high level interface that will check
if the codebase inside a GitHub repository contains the
dependency management files for a given platform. If it does,
then it will parse those files and return a list of dependencies
that will be resolved by the DependencyResolver.

For instance, https://github.com/toptal/platform is a Ruby
project and because of that it contains a Gemfile. The
RubyRepositoryFilter will parse the Gemfile.