Add support for Node.js platform and improve architecture
Node.js is now fully supported. However, only the first-level
'package.json' will be matched and parsed. Repositories with
multiple 'package.json' files are not supported at the moment.

Renamed a few domain classes:
    - RepositoryFilter -> RepositoryMatcher
    - PlatformResolver -> Platform

The result of a repository matching is now a list of
"package descriptors". Each package descriptor contains
the reference to the original repository as well as
development and runtime dependencies.

Renamed a few VCR tapes.

Tiny improvements over the codebase.