Don't abort dependency resolution when package version is not found
When resolving dependencies for the Node.js platform I've found
that many dependencies had at least one version not found in
their resolution tree.

The previous implementation would throw an exception and the
resolution would be aborted for that dependency.

This new implementation continues the dependency resolution,
marking in the tree that a version was not found.

Also, improves the main script and console output.