Ability to batch process all repositories in an organization
This result in a tremendous performance gain, since the dependencies are
cached by license cop.

For example, the processing of Ruby dependencies for all 165 repositories
in  this organization was reduced from several hours to 15 minutes.

This is also critical for processing Node.js dependencies, since they take
alone hours for a single repository.